pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "local-ai"
version = "0.2.0"
description = "Локальный Telegram-бот и сервисы вокруг Ollama (router, OCR, vision, eval)."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "requests>=2.31.0",
    "python-telegram-bot>=21.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "pymupdf>=1.24.0",
    "Pillow>=10.0.0",
    "numpy>=1.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "httpx>=0.27.0",
]
ssh = [
    "paramiko>=3.4.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
